Extract timezone related methods.
Also move a few more modifiers I missed the first time around.
